# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: Band Manager API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /account/onboarding:
    post:
      operationId: onboardOwner
      summary: Create the verified user's owner account and band
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardOwnerRequest'
      responses:
        '201':
          description: Owner account and band created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentAccountResponse'
        '400':
          description: Invalid onboarding request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Account or band already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Email is not verified or provider identity does not match
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Identity provider could not inspect the verified user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /me:
    get:
      operationId: getCurrentAccount
      summary: Get the authenticated user and active band
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current account context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentAccountResponse'
        '401':
          description: Missing or invalid session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Authenticated user has not completed onboarding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /financial-reports:
    get:
      operationId: getFinancialReport
      summary: Get finalized merch sales financial report
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Inclusive local band date in YYYY-MM-DD format.
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Inclusive local band date in YYYY-MM-DD format.
      responses:
        '200':
          description: Financial report for finalized merch sales
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialReportResponse'
        '400':
          description: Invalid report date range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Role cannot read reports in alpha
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Report query failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
      description: Unique key scoped by band and operation for mutating requests.
  schemas:
...
    OnboardOwnerRequest:
      type: object
      required:
        - bandName
        - bandTimezone
      properties:
        bandName:
          type: string
          minLength: 1
        bandTimezone:
          type: string
          minLength: 1
          example: America/Recife
    CurrentAccountResponse:
      type: object
      required:
        - user
        - activeBand
      properties:
        user:
          $ref: '#/components/schemas/User'
        activeBand:
          $ref: '#/components/schemas/BandMembership'
    User:
      type: object
      required:
        - id
        - email
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
    BandMembership:
      type: object
      required:
        - bandId
        - bandName
        - role
        - canWrite
      properties:
        bandId:
          type: string
          format: uuid
        bandName:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        canWrite:
          type: boolean
...
    Role:
      type: string
      enum:
        - owner
        - admin
        - member
        - viewer
...
    Money:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: integer
          minimum: 0
        currency:
          type: string
          enum:
            - BRL
    SignedMoney:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: integer
        currency:
          type: string
          enum:
            - BRL
    FinancialReportResponse:
      type: object
      required:
        - range
        - summary
        - paymentMethods
        - categories
        - products
        - days
      properties:
        range:
          $ref: '#/components/schemas/FinancialReportRange'
        summary:
          $ref: '#/components/schemas/FinancialReportSummary'
        paymentMethods:
          type: array
          items:
            $ref: '#/components/schemas/FinancialReportPaymentMethodBreakdown'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/FinancialReportCategoryBreakdown'
        products:
          type: array
          items:
            $ref: '#/components/schemas/FinancialReportProductBreakdown'
        days:
          type: array
          items:
            $ref: '#/components/schemas/FinancialReportDayBreakdown'
    FinancialReportRange:
      type: object
      required:
        - from
        - to
        - timezone
      properties:
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        timezone:
          type: string
          minLength: 1
    FinancialReportSummary:
      type: object
      required:
        - saleCount
        - itemCount
        - grossRevenue
        - totalHistoricalCost
        - expectedProfit
      properties:
        saleCount:
          type: integer
          minimum: 0
        itemCount:
          type: integer
          minimum: 0
        grossRevenue:
          $ref: '#/components/schemas/Money'
        totalHistoricalCost:
          $ref: '#/components/schemas/Money'
        expectedProfit:
          $ref: '#/components/schemas/SignedMoney'
    FinancialReportPaymentMethodBreakdown:
      allOf:
        - $ref: '#/components/schemas/FinancialReportAggregate'
        - type: object
          required:
            - method
          properties:
            method:
              type: string
              enum:
                - cash
                - pix
                - card
    FinancialReportCategoryBreakdown:
      allOf:
        - $ref: '#/components/schemas/FinancialReportAggregate'
        - type: object
          required:
            - category
          properties:
            category:
              $ref: '#/components/schemas/InventoryCategory'
    FinancialReportProductBreakdown:
      allOf:
        - $ref: '#/components/schemas/FinancialReportAggregate'
        - type: object
          required:
            - productId
            - productName
            - category
          properties:
            productId:
              type: string
              format: uuid
            productName:
              type: string
            category:
              $ref: '#/components/schemas/InventoryCategory'
    FinancialReportDayBreakdown:
      allOf:
        - $ref: '#/components/schemas/FinancialReportAggregate'
        - type: object
          required:
            - date
          properties:
            date:
              type: string
              format: date
    FinancialReportAggregate:
      type: object
      required:
        - saleCount
        - itemCount
        - grossRevenue
        - totalHistoricalCost
        - expectedProfit
      properties:
        saleCount:
          type: integer
          minimum: 0
        itemCount:
          type: integer
          minimum: 0
        grossRevenue:
          $ref: '#/components/schemas/Money'
        totalHistoricalCost:
          $ref: '#/components/schemas/Money'
        expectedProfit:
          $ref: '#/components/schemas/SignedMoney'
...
    InventoryCategory:
      type: string
      enum:
        - shirt
        - hoodie
        - tote_bag
        - patch
        - sticker
        - vinyl
        - cd
        - cassette
        - accessory
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string